requirements.txt
absl-py==1.4.0
appdirs==1.4.4
array-record==0.2.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
deeplookup==0.1.0
dm-tree==0.1.8
dnspython==2.3.0
docker-pycreds==0.4.0
etils==1.3.0
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
jax==0.4.10
joblib==1.2.0
keract==4.5.1
keras==2.12.0
keras-rl2==1.0.5
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.40.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
nltk==3.8.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
promise==2.3
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyts==0.12.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.3
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
